requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.34.151
botocore==1.34.151
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==64.0.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
jmespath==1.0.1
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
stripe==9.8.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.0
urllib3==2.2.1
watchdog==4.0.2
Werkzeug==3.0.3
whitenoise==6.7.0
zope.interface==6.4.post2
